eme_playback/test_eme_unset_mediakeys_and_capture.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Encrypted Media Extensions</title>
  <script type="text/javascript" src="SimpleTest.js"></script>
  <script type="text/javascript" src="manifest.js"></script>
  <script type="text/javascript" src="eme.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

function startTest(test, token)
{
  var sessions = [];
  function onSessionCreated(session) {
    sessions.push(session);
  }

  let v = document.createElement("video");
  document.body.appendChild(v);

  let finish = new EMEPromise;

  function onVideoEnded(ev) {
    console.log(" got ended event ");
    v.removeEventListener("ended", onVideoEnded);

    function playNewVideo() {
      var p1 = once(v, 'ended', function onended(e) {
        console.log(" bipbop.mp4 playback ended !!");
      });
      var p2 = once(v, 'playing', function onplay(e) {
        console.log(" bipbop is playing .... ");
        canvasElem = document.createElement('canvas');
        document.body.appendChild(canvasElem);
        ctx2d = canvasElem.getContext('2d');
        ctx2d.drawImage(v, 0, 0);  
      });
      v.src = 'bipbop.mp4';
      v.play();
      Promise.all([p1, p2]).then(() => {
        console.log(" Test finish !!");
      });
    }

    Promise.all(sessions.map(s => s.close()))
    .then(() => {
      /* load() is invoked when going to YT homepage by click YT logo.
       * Simulating the scenario here.
       */
      // v.load();
      v.setMediaKeys(null)
      .then(() => {
        console.log(" v.setMediaKeys OK !!, try to play clear content");
        playNewVideo();
      }, () => {
        console.log(" v.setMediaKeys NG !!, still try to play clear content");
        playNewVideo();
      });;
    });


  }

  v.addEventListener("ended", onVideoEnded);

  Promise.all([
    LoadInitData(v, test, token),
    CreateAndSetMediaKeys(v, test, token),
    LoadTest(test, v, token)])
  .then(values => {
    v.play();
    let initData = values[0];
    initData.map(ev => {
      let session = v.mediaKeys.createSession();
      onSessionCreated(session);
      MakeRequest(test, token, ev, session);
    });
    console.log(" start to play ");
  })
  .then(() => {
    console.log(" finish.promise ");
    return finish.promise;
  })
  .catch(reason => ok(false, reason))
}

startTest(gEMETests[0], 'TEST >>> ');

</script>
</pre>
</body>
</html>
